backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.2.1
httpx==0.26.0
websockets==12.0
python-dotenv==1.0.1
tenacity==8.2.3
